backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
Pillow
requests
httpx
